refactor(auth): move api_client to its own module and create it at the app level
