+++ src/app.rs
@@ -1,8 +1,12 @@
-use crate::{auth, prelude::*, run};
+use crate::{api_client::CodSpeedAPIClient, auth, prelude::*, run};
 use clap::{Parser, Subcommand};
 
 #[derive(Parser, Debug)]
-struct Cli {
+pub struct Cli {
+    /// The URL of the CodSpeed GraphQL API
+    #[arg(long, env = "CODSPEED_API_URL", global = true, hide = true)]
+    pub api_url: Option<String>,
+
     #[command(subcommand)]
     command: Commands,
 }
@@ -17,10 +21,11 @@ enum Commands {
 
 pub async fn run() -> Result<()> {
     let cli = Cli::parse();
+    let api_client = CodSpeedAPIClient::from(&cli);
 
     match cli.command {
         Commands::Run(args) => run::run(args).await?,
-        Commands::Auth(args) => auth::run(args).await?,
+        Commands::Auth(args) => auth::run(args, &api_client).await?,
     }
     Ok(())
 }
