+++ src/auth.rs
@@ -1,19 +1,13 @@
 use std::time::Duration;
 
-use crate::{config::Config, logger::get_local_logger, prelude::*};
+use crate::logger::get_local_logger;
+use crate::{api_client::CodSpeedAPIClient, config::Config, prelude::*};
 use clap::{Args, Subcommand};
-use gql_client::{Client as GQLClient, ClientConfig};
-use nestify::nest;
-use serde::{Deserialize, Serialize};
 use simplelog::CombinedLogger;
 use tokio::time::{sleep, Instant};
 
 #[derive(Debug, Args)]
 pub struct AuthArgs {
-    /// The URL of the CodSpeed GraphQL API
-    #[arg(long, env = "CODSPEED_API_URL", global = true, hide = true)]
-    api_url: Option<String>,
-
     #[command(subcommand)]
     command: AuthCommands,
 }
@@ -31,9 +25,8 @@ fn init_logger() -> Result<()> {
     Ok(())
 }
 
-pub async fn run(args: AuthArgs) -> Result<()> {
+pub async fn run(args: AuthArgs, api_client: &CodSpeedAPIClient) -> Result<()> {
     init_logger()?;
-    let api_client = CodSpeedAPIClient::from(&args);
 
     match args.command {
         AuthCommands::Login => login(api_client).await?,
@@ -41,90 +34,9 @@ pub async fn run(args: AuthArgs) -> Result<()> {
     Ok(())
 }
 
-nest! {
-    #[derive(Debug, Deserialize, Serialize)]*
-    #[serde(rename_all = "camelCase")]*
-    struct CreateLoginSessionData {
-        create_login_session: struct CreateLoginSessionPayload {
-            callback_url: String,
-            session_id: String,
-        }
-    }
-}
-
-nest! {
-    #[derive(Debug, Deserialize, Serialize)]*
-    #[serde(rename_all = "camelCase")]*
-    struct ConsumeLoginSessionData {
-        consume_login_session: struct ConsumeLoginSessionPayload {
-            token: Option<String>
-        }
-    }
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ConsumeLoginSessionVars {
-    session_id: String,
-}
-
-struct CodSpeedAPIClient {
-    gql_client: GQLClient,
-}
-
-impl From<&AuthArgs> for CodSpeedAPIClient {
-    fn from(args: &AuthArgs) -> Self {
-        Self {
-            gql_client: build_gql_api_client(args.api_url.clone()),
-        }
-    }
-}
-
-const CODSPEED_GRAPHQL_ENDPOINT: &str = "https://gql.codspeed.io/";
-
-fn build_gql_api_client(api_url: Option<String>) -> GQLClient {
-    let endpoint = api_url.unwrap_or_else(|| CODSPEED_GRAPHQL_ENDPOINT.to_string());
-
-    GQLClient::new_with_config(ClientConfig {
-        endpoint,
-        timeout: Some(10),
-        headers: Default::default(),
-        proxy: None,
-    })
-}
-
-impl CodSpeedAPIClient {
-    async fn create_login_session(&self) -> Result<CreateLoginSessionPayload> {
-        let response = self
-            .gql_client
-            .query_unwrap::<CreateLoginSessionData>(include_str!("queries/CreateLoginSession.gql"))
-            .await;
-        match response {
-            Ok(response) => Ok(response.create_login_session),
-            Err(err) => bail!("Failed to create login session: {}", err),
-        }
-    }
-
-    async fn consume_login_session(&self, session_id: &str) -> Result<ConsumeLoginSessionPayload> {
-        let response = self
-            .gql_client
-            .query_with_vars_unwrap::<ConsumeLoginSessionData, ConsumeLoginSessionVars>(
-                include_str!("queries/ConsumeLoginSession.gql"),
-                ConsumeLoginSessionVars {
-                    session_id: session_id.to_string(),
-                },
-            )
-            .await;
-        match response {
-            Ok(response) => Ok(response.consume_login_session),
-            Err(err) => bail!("Failed to use login session: {}", err),
-        }
-    }
-}
-
 const LOGIN_SESSION_MAX_DURATION: Duration = Duration::from_secs(60 * 5); // 5 minutes
 
-async fn login(api_client: CodSpeedAPIClient) -> Result<()> {
+async fn login(api_client: &CodSpeedAPIClient) -> Result<()> {
     debug!("Login to CodSpeed");
     debug!("Creating login session...");
     let login_session_payload = api_client.create_login_session().await?;
